+++ sources/ippcp/cpinit.c
@@ -196,8 +196,8 @@ static int cpGetFeatures( Ipp64u* pFeaturesMask )
       cpGetReg( (int*)buf, 0x7, 1 ); // Intel® SM3, SM4, and SHA512 instructions set check
       eax_ = (Ipp32u)buf[0];
       if (eax_ & BIT00) mask |= ippCPUID_AVX2SHA512;  // test bit eax[00]
-      if (eax_ & BIT01) mask |= ippCPUID_AVX2SM3;  // test bit eax[01]
-      if (eax_ & BIT02) mask |= ippCPUID_AVX2SM4;  // test bit eax[02]
+      if (eax_ & BIT01) mask |= ippCPUID_AVX2SM3;     // test bit eax[01]
+      if (eax_ & BIT02) mask |= ippCPUID_AVX2SM4;     // test bit eax[02]
 
 
        if (mask & ippCPUID_AVX512F) {