+++ sources/ippcp/hash/sha256/pcphashmethodset_sha256_ni.c
@@ -35,7 +35,8 @@
 /*F*
 //    Name: ippsHashMethodSet_SHA256_NI
 //
-// Purpose: Setup SHA256 method (using the Intel® Secure Hash Algorithm - New Instructions (Intel® SHA-NI) instruction set).
+// Purpose: Setup SHA256 method
+// (using the Intel® Secure Hash Algorithm - New Instructions (Intel® SHA-NI) instruction set).
 //
 // Returns:                Reason:
 //    ippStsNullPtrErr           pMethod == NULL