+++ sources/ippcp/hash/sha256/pcphashstatemethodset_sha256_ni.c
@@ -35,7 +35,8 @@
 /*F*
 //    Name: ippsHashStateMethodSet_SHA256_NI
 //
-// Purpose: Setup SHA256 method inside the hash state (using the Intel® Secure Hash Algorithm - New Instructions (Intel® SHA-NI) instruction set).
+// Purpose: Setup SHA256 method inside the hash state
+// (using the Intel® Secure Hash Algorithm - New Instructions (Intel® SHA-NI) instruction set).
 //
 // Returns:                Reason:
 //    ippStsNullPtrErr           pMethod == NULL or pState == NULL