+++ sources/ippcp/hash/sha512/pcphashmethodset_sha512.c
@@ -14,12 +14,12 @@
 * limitations under the License.
 *************************************************************************/
 
-/* 
-// 
+/*
+//
 //  Purpose:
 //     Cryptography Primitive.
 //     SHA512 message digest
-// 
+//
 //  Contents:
 //        ippsHashMethodSet_SHA512()
 //