+++ sources/ippcp/hash/sha512/pcphashmethodset_sha512_224_tt.c
@@ -1,5 +1,5 @@
 /*************************************************************************
-* Copyright (C) 2022 Intel Corporation
+* Copyright (C) 2024 Intel Corporation
 *
 * Licensed under the Apache License,  Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -18,10 +18,10 @@
 //
 //  Purpose:
 //     Cryptography Primitive.
-//     SHA512/256 message digest
+//     SHA512/224 message digest
 //
 //  Contents:
-//        ippsHashStateMethodSet_SHA512_256()
+//        ippsHashMethodSet_SHA512_224_TT()
 //
 */
 
@@ -33,31 +33,35 @@
 #include "hash/sha512/pcpsha512stuff.h"
 
 /*F*
-//    Name: ippsHashStateMethodSet_SHA512_256
+//    Name: ippsHashMethodSet_SHA512_224_TT
 //
-// Purpose: Setup SHA512/256 method inside the hash state.
+// Purpose: Setup SHA512/224 method
+// (using the Intel® SHA512 instruction set if it is available at run time)
 //
 // Returns:                Reason:
-//    ippStsNullPtrErr        pMethod == NULL or pState == NULL
+//    ippStsNullPtrErr        pMethod == NULL
 //    ippStsNoErr             no errors
 //
 *F*/
 
-IPPFUN( IppStatus, ippsHashStateMethodSet_SHA512_256, (IppsHashState_rmf* pState, IppsHashMethod* pMethod) )
+IPPFUN( IppStatus, ippsHashMethodSet_SHA512_224_TT, (IppsHashMethod* pMethod) )
 {
    /* test pointers */
-   IPP_BAD_PTR2_RET(pState, pMethod);
+   IPP_BAD_PTR1_RET(pMethod);
 
-   HASH_METHOD(pState) = pMethod;
-
-   pMethod->hashAlgId     = ippHashAlg_SHA512_256;
-   pMethod->hashLen       = IPP_SHA256_DIGEST_BITSIZE/8;
+   pMethod->hashAlgId     = ippHashAlg_SHA512_224;
+   pMethod->hashLen       = IPP_SHA224_DIGEST_BITSIZE/8;
    pMethod->msgBlkSize    = MBS_SHA512;
    pMethod->msgLenRepSize = MLR_SHA512;
-   pMethod->hashInit      = sha512_256_hashInit;
+   pMethod->hashInit      = sha512_224_hashInit;
    pMethod->hashUpdate    = sha512_hashUpdate;
-   pMethod->hashOctStr    = sha512_256_hashOctString;
+   pMethod->hashOctStr    = sha512_224_hashOctString;
    pMethod->msgLenRep     = sha512_msgRep;
+
+#if (_SHA512_ENABLING_==_FEATURE_TICKTOCK_ || _SHA512_ENABLING_==_FEATURE_ON_)
+   if(IsFeatureEnabled(ippCPUID_AVX2SHA512))
+      pMethod->hashUpdate = sha512_hashUpdate_ni;
+#endif
 
    return ippStsNoErr;
 }